// types/Receiver/ReceiverInfomation.ts
export type receiverInfomation = {
  id: string
  shipItem: string
  receiverName: string
  receiverTelNumber: string
  receiverAddress: string
  reservationStatus: string
  billNumber: string
  plannedCollectPlace: string
  receptionDate: string
  expectedCollectionDate: string
  collectionDate: string
  reservationClassification: string
  reservationNumber: string
  senderId: string
  senderName: string
  sendedOrganizationName: string
  senderTelNumber: string
  senderAddress: string
  quantity: string
  receptionNumber: string
  reservatedMedia: string
}